@@ -1,6 +1,6 @@
 {
   "name": "tridoron",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "description": "Router library for React",
   "main": "lib/index.js",
   "repository": "https://github.com/ushiboy/tridoron",
@@ -18,35 +18,40 @@
   "author": "ushiboy",
   "license": "MIT",
   "devDependencies": {
-    "babel-cli": "^6.16.0",
-    "babel-loader": "^6.2.5",
-    "babel-preset-es2015": "^6.14.0",
+    "babel-cli": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-preset-es2015": "^6.24.1",
     "babel-preset-power-assert": "^1.0.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-register": "^6.16.3",
-    "connect": "^3.5.0",
-    "connect-history-api-fallback": "^1.3.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "connect": "^3.6.5",
+    "connect-history-api-fallback": "^1.5.0",
     "connect-livereload": "^0.6.0",
-    "enzyme": "^2.5.1",
-    "eslint": "^3.6.1",
-    "eslint-plugin-react": "^6.3.0",
+    "enzyme": "^3.1.1",
+    "enzyme-adapter-react-16": "^1.0.4",
+    "eslint": "^4.11.0",
+    "eslint-plugin-react": "^7.4.0",
     "gulp": "^3.9.1",
     "gulp-livereload": "^3.8.1",
-    "gulp-load-plugins": "^1.3.0",
-    "gulp-util": "^3.0.7",
-    "jsdom": "^9.6.0",
-    "mocha": "^3.1.0",
-    "power-assert": "^1.4.1",
-    "react-addons-test-utils": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "serve-static": "^1.11.1",
-    "sinon": "^1.17.6",
-    "webpack": "^1.13.2"
+    "gulp-load-plugins": "^1.5.0",
+    "gulp-util": "^3.0.8",
+    "jsdom": "^11.3.0",
+    "mocha": "^4.0.1",
+    "power-assert": "^1.4.4",
+    "react": "^16.1.1",
+    "react-addons-test-utils": "^15.6.2",
+    "react-dom": "^16.1.1",
+    "serve-static": "^1.13.1",
+    "sinon": "^4.1.2",
+    "webpack": "^3.8.1"
   },
   "dependencies": {
     "fbemitter": "^2.1.1",
-    "path-to-regexp": "^1.5.3",
-    "react": "^15.3.2",
-    "url-parse": "^1.1.4"
+    "path-to-regexp": "^2.1.0",
+    "prop-types": "^15.6.0",
+    "url-parse": "^1.2.0"
+  },
+  "peerDependencies": {
+    "react": "^15.0.0-0 || ^16.0.0-0"
   }
 }
