@@ -21,11 +21,11 @@ const gulp = require('gulp'),
       filename: '[name]-bundle.js'
     },
     module: {
-      loaders: [
+      rules: [
         {
           test: /\.js$/,
           exclude: /node_modules/,
-          loader: 'babel'
+          loader: 'babel-loader'
         }
       ]
     }
